testing and individual formulas
