+++ policyengine_us/tests/policy/baseline/gov/states/ca/cdss/state_supplement/ca_state_supplement_medical_care_facility_amount.yaml
@@ -1,0 +1,35 @@
+- name: One person in medical care facility
+  period: 2024
+  input:
+    people:
+      person1:
+        ca_in_medical_care_facility: true
+      person2:
+        ca_in_medical_care_facility: false
+    spm_units:
+      spm_unit:
+        members: [person1, person2]
+    households:
+      household:
+        members: [person1, person2]
+        state_code: CA
+  output:
+    ca_state_supplement_medical_care_facility_amount: 504
+
+- name: Nobody in medical care facility
+  period: 2024
+  input:
+    people:
+      person1:
+        ca_in_medical_care_facility: false
+      person2:
+        ca_in_medical_care_facility: false
+    spm_units:
+      spm_unit:
+        members: [person1, person2]
+    households:
+      household:
+        members: [person1, person2]
+        state_code: CA
+  output:
+    ca_state_supplement_medical_care_facility_amount: 0
